requirements.txt
flask
gunicorn
torch
torchvision
albumentations
opencv-python-headless
pillow
numpy
